fix fatal trying to do a sub between an int and a string

+++ htdocs/compta/facture/card.php
@@ -2742,8 +2742,7 @@ if (empty($reshook)) {
 		// Invoice situation
 		if (getDolGlobalInt('INVOICE_USE_SITUATION') == 2) {
 			$previousprogress = $line->get_allprev_progress($line->fk_facture);
-			$fullprogress = price2num(GETPOST('progress', 'alpha'));
-
+			$fullprogress = (int)price2num(GETPOST('progress', 'alpha'));
 			if ($fullprogress < $previousprogress) {
 				$error++;
 				setEventMessages($langs->trans('CantBeLessThanMinPercent'), null, 'errors');
